Add a contributing guideline for this project
I've written a new contributing guideline for this project that reflects
the higher standards that we will have for contributions to the API
project. In short, the contributing guidelines explain that this
project is essential to our internal infrastructure and needs to be
managed as such.

It also refers to our Code of Conduct and the license that we've chosen
for this project.