Add CODE_OF_CONDUCT.md for GitHub's interface
GitHub will automatically reference a CODE_OF_CONDUCT.md file if it is
present in the root of the repository. That's why I've added a stub file
that refers the reader to the Code of Conduct page on our website. This
means that we can have GitHub refer to our CoC without having to
maintain it in two different places.