Add initial README.md to the project
I've added an initial README.md to the project that will be updated as
the project evolves. As of now, it contains a short description of the
project and explains how to contribute to the project.

Once the docker containers are set up, I will add information on how to
run a Docker-based development set-up locally.