Add pydantic and fastapi to the dependencies
I've added both pydantic and fastapi to the project's main dependencies,
as they are both needed to run the project. I have not added unicorn to
the main project dependencies, as locking the main dependencies on
Windows would create a lockfile without uvloop.

Instead, uvicorn will be installed in the builder stage of building the
Docker image, where an appropriate build environment (not Windows) can
be guaranteed.