Fallback to the old API for undefined routes.
This adds a new endpoint handler using the `path` converter in order to match
requests directed to any subtree. As long as this endpoint is the last to be
included in the router, all requests not matching defined routes under the
`/api` endpoint will be forwarded to the old API.

Two new settings are added for this. One is the endpoint of the old API to
forward to, which may change in the future due to the pending subdomain removal
pull request. The other is the API token to use for the new API authenticating
with the old API: Whilst everything else sent by the client is forwarded to the
old API, the authorization schemes between the new and old API differ.

The entire point behind this functionality is to allow us to incrementally port
over endpoints to the new API, whilst reconfiguring clients to use the new API
completely.

Closes #8.